backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.3
redis[asyncio]==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
aiosmtplib==3.0.1
jinja2==3.1.3
shapely==2.0.2
pyproj==3.6.1
reportlab==4.0.9
openpyxl==3.1.2
websockets==12.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
